Renamed ctsLilypond.py -> lilypond.py
