+++ examples/lilypondClipExample.py
@@ -2,7 +2,7 @@ import os
 import subprocess
 
 from chiptunesak import midi
-from chiptunesak import ctsLilypond
+from chiptunesak.lilypond import Lilypond
 from chiptunesak.constants import project_to_absolute_path
 
 """
@@ -27,7 +27,7 @@ chirp_song = midi.MIDI().to_chirp(input_file, quantization='16', polyphony=False
 mchirp_song = chirp_song.to_mchirp()
 
 # Create the lilpond I/O class
-lp = ctsLilypond.Lilypond()
+lp = Lilypond()
 # Set the format to do a clip and set the measures to those you want
 lp.set_options(format='clip', measures=mchirp_song.tracks[0].measures[3:8])
 # Write it straight to a file
